+++ contentful_management/array.py
@@ -6,7 +6,7 @@ This module implements the Array class.
 
 API reference: https://www.contentful.com/developers/docs/references/content-management-api/#/introduction/collection-resources-and-pagination
 
-:copyright: (c) 2017 by Contentful GmbH.
+:copyright: (c) 2018 by Contentful GmbH.
 :license: MIT, see LICENSE for more details.
 """