+++ contentful_management/content_type_field_validation.py
@@ -9,7 +9,7 @@ This module implements the ContentTypeFieldValidation class.
 
 API reference: https://www.contentful.com/developers/docs/references/content-management-api/#/reference/content-types/content-type
 
-:copyright: (c) 2017 by Contentful GmbH.
+:copyright: (c) 2018 by Contentful GmbH.
 :license: MIT, see LICENSE for more details.
 """