+++ contentful_management/entry.py
@@ -11,7 +11,7 @@ This module implements the Entry class.
 
 API reference: https://www.contentful.com/developers/docs/references/content-delivery-api/#/reference/entries
 
-:copyright: (c) 2017 by Contentful GmbH.
+:copyright: (c) 2018 by Contentful GmbH.
 :license: MIT, see LICENSE for more details.
 """