+++ contentful_management/environment.py
@@ -13,7 +13,7 @@ This module implements the Environment class.
 
 API reference: https://www.contentful.com/developers/docs/references/content-management-api/#/reference/locales
 
-:copyright: (c) 2017 by Contentful GmbH.
+:copyright: (c) 2018 by Contentful GmbH.
 :license: MIT, see LICENSE for more details.
 """