+++ contentful_management/errors.py
@@ -6,7 +6,7 @@ This module implements the Error classes.
 
 API reference: https://www.contentful.com/developers/docs/references/content-delivery-api/#/introduction/errors
 
-:copyright: (c) 2017 by Contentful GmbH.
+:copyright: (c) 2018 by Contentful GmbH.
 :license: MIT, see LICENSE for more details.
 """