+++ contentful_management/organizations_proxy.py
@@ -10,7 +10,7 @@ This module implements the OrganizationsProxy class.
 
 API reference: https://www.contentful.com/developers/docs/references/content-management-api/#/reference/organizations
 
-:copyright: (c) 2017 by Contentful GmbH.
+:copyright: (c) 2018 by Contentful GmbH.
 :license: MIT, see LICENSE for more details.
 """