+++ contentful_management/role.py
@@ -9,7 +9,7 @@ This module implements the Role class.
 
 API reference: https://www.contentful.com/developers/docs/references/content-management-api/#/reference/roles
 
-:copyright: (c) 2017 by Contentful GmbH.
+:copyright: (c) 2018 by Contentful GmbH.
 :license: MIT, see LICENSE for more details.
 """