+++ contentful_management/snapshot.py
@@ -9,7 +9,7 @@ This module implements the Snapshot class.
 
 API reference: https://www.contentful.com/developers/docs/references/content-management-api/#/reference/snapshots
 
-:copyright: (c) 2017 by Contentful GmbH.
+:copyright: (c) 2018 by Contentful GmbH.
 :license: MIT, see LICENSE for more details.
 """