+++ contentful_management/upload.py
@@ -9,7 +9,7 @@ This module implements the Upload class.
 
 API reference: https://www.contentful.com/developers/docs/references/content-management-api/#/reference/uploads
 
-:copyright: (c) 2017 by Contentful GmbH.
+:copyright: (c) 2018 by Contentful GmbH.
 :license: MIT, see LICENSE for more details.
 """