+++ contentful_management/webhook_health.py
@@ -9,7 +9,7 @@ This module implements the WebhookHealth class.
 
 API reference: https://www.contentful.com/developers/docs/references/content-management-api/#/reference/webhook-calls/webhook-health
 
-:copyright: (c) 2017 by Contentful GmbH.
+:copyright: (c) 2018 by Contentful GmbH.
 :license: MIT, see LICENSE for more details.
 """